Move ResPrototype to where it is used
