+++ src/ert/__init__.py
@@ -1,3 +1,6 @@
+"""
+Ert - Ensemble Reservoir Tool - a package for reservoir modeling.
+"""
 from .config import ErtScript
 from .data import MeasuredData
 from .job_queue import JobStatusType
